cronjob/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.33
protobuf==3.20.3
psycopg2==2.9.7
pydantic==2.0.2
pydantic_core==2.1.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
